+++ src/gui/AppStartMenu.java
@@ -16,7 +16,11 @@ import java.sql.SQLException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.swing.JOptionPane;
+import lanword.interfaces.bd.BDManagment;
 import lanword.interfaces.bd.BDResolver;
+import lanword.interfaces.bd.IBDGestionPalabras;
+import lanword.modelo.Idioma;
+import lanword.modelo.Palabra;
 
 /**
  *
@@ -159,9 +163,21 @@ public class AppStartMenu extends javax.swing.JFrame {
                     
                     if (checkIfAnotherLanwordRuns())
                         JOptionPane.showMessageDialog(null, "Lanword ya se está ejecutando.");
-                    else 
+                    else {
+
+                        if (BDResolver.getInstance().idiomas.buscar().isEmpty()) {
+                            int choice = 
+                                    JOptionPane.showConfirmDialog(null, "No hay palabras en la base de datos, \n"
+                                            + "¿quieres añadir las palabras predeterminadas?", "Palabras predeterminadas",
+                                            JOptionPane.YES_NO_OPTION);
+
+                            if (choice == JOptionPane.YES_OPTION) {
+                                generarPalabras();
+                            }
+                        }
                         new AppStartMenu().setVisible(true);
-                    
+                    }
+
                 } catch (SQLException ex) {
                     Logger.getLogger(AppStartMenu.class.getName()).log(Level.SEVERE, null, ex);
                 } catch (ClassNotFoundException ex) {
@@ -183,6 +199,72 @@ public class AppStartMenu extends javax.swing.JFrame {
             init.createNewFile();
         
         return otherLanwordRuns;
+    };
+    
+    private static void generarPalabras() throws SQLException, ClassNotFoundException, IOException {
+        IBDGestionPalabras bd = BDResolver.getInstance().palabras;
+        Idioma espanol = new Idioma("Español");
+        Idioma ingles = new Idioma("Inglés");
+        
+        Palabra laptop = new Palabra("Laptop", ingles);
+        Palabra pc = new Palabra("Personal Computer", ingles);
+        Palabra hub = new Palabra("Hub", ingles);
+        Palabra router = new Palabra("Router", ingles);
+        Palabra monitor = new Palabra("Monitor", ingles);
+        Palabra mouse = new Palabra("Mouse", ingles);
+        Palabra keyboard = new Palabra("Keyboard", ingles);
+        Palabra motherboard = new Palabra("Motherboard", ingles);
+        Palabra harddisk = new Palabra("Hard disk", ingles);
+        Palabra opticaldrive = new Palabra("Optical drive", ingles);
+        Palabra printer = new Palabra("Printer", ingles);
+        Palabra codification = new Palabra("Codificacion", ingles);
+        Palabra software = new Palabra("Software", ingles);
+        Palabra hardware = new Palabra("hardware", ingles);
+        
+        Palabra portatil = new Palabra("Portátil", espanol);
+        Palabra pc_es = new Palabra("Ordenador personal", espanol);
+        Palabra repetidor = new Palabra("Repetidor", espanol);
+        Palabra encaminador = new Palabra("Encaminador", espanol);
+        Palabra monitor_es = new Palabra("Monitor", espanol);
+        Palabra raton = new Palabra("Ratón", espanol);
+        Palabra teclado = new Palabra("Teclado", espanol);
+        Palabra placabase = new Palabra("Placa base", espanol);
+        Palabra discoduro = new Palabra("Disco duro", espanol);
+        Palabra lectoroptico = new Palabra("Lector óptico", espanol);
+        Palabra impresora = new Palabra("Impresora", espanol);
+        Palabra codificacion = new Palabra("Codificación", espanol);
+        Palabra software_es = new Palabra("Software", espanol);
+        Palabra hardware_es = new Palabra("Hardware", espanol);
+        
+        laptop.anyadirTraduccion(portatil);
+        pc.anyadirTraduccion(pc_es);
+        hub.anyadirTraduccion(repetidor);
+        router.anyadirTraduccion(encaminador);
+        monitor.anyadirTraduccion(monitor_es);
+        mouse.anyadirTraduccion(raton);
+        keyboard.anyadirTraduccion(teclado);
+        motherboard.anyadirTraduccion(placabase);
+        harddisk.anyadirTraduccion(discoduro);
+        opticaldrive.anyadirTraduccion(lectoroptico);
+        printer.anyadirTraduccion(impresora);
+        codification.anyadirTraduccion(codificacion);
+        software.anyadirTraduccion(software_es);
+        hardware.anyadirTraduccion(hardware_es);
+        
+        bd.anyadir(laptop);
+        bd.anyadir(pc);
+        bd.anyadir(hub);
+        bd.anyadir(router);
+        bd.anyadir(monitor);
+        bd.anyadir(mouse);
+        bd.anyadir(keyboard);
+        bd.anyadir(motherboard);
+        bd.anyadir(harddisk);
+        bd.anyadir(opticaldrive);
+        bd.anyadir(printer);
+        bd.anyadir(codification);
+        bd.anyadir(software);
+        bd.anyadir(hardware);
     };
 
     // Variables declaration - do not modify//GEN-BEGIN:variables