+++ CHANGES.md
@@ -4,7 +4,7 @@
  - [ ] Solucionar el bug de las traducciones transitivas. Issue #6.
  - [ ] Solucionar el bug gráfico de la tabla de palabras. Issue #5.
 - *Funcionalidad*
- - [ ] Permitir cargar una estructura de palabras si la base de datos está vacía.
+ - [x] Permitir cargar una estructura de palabras si la base de datos está vacía.
  - [ ] Añadir texto informativo acerca de como funciona el juego.
  - [ ] No permitir iniciar el juego si no hay al menos dos idiomas.
  - [ ] Permitir desde administración ir a juego y vicebersa.